Revert "Try to fix failed tests"
This reverts commit fd3dd1d6c1cd04e2a74a9bf4ee518a5e2a3a1aa0.

+++ Embedded/Arduino/Motor_controller/src/Calculations/Drive.cpp
@@ -71,11 +71,11 @@ DriveState Drive::GetState() {
    Om state är STOPPED, ska en annan funktion anropas som stoppar motorerna
 */
 void Drive::ExecuteDriveLogic() {
-    int currentSpeed = round(CalculateHypotenuse());
-    int leftSpeed = round(CalculateLeftSpeedFunc());
-    int rightSpeed = round(CalculateRightSpeedFunc());
+    float currentSpeed = round(CalculateHypotenuse());
+    float leftSpeed = round(CalculateLeftSpeedFunc());
+    float rightSpeed = round(CalculateRightSpeedFunc());
     DriveState dir = GetState();
-    printf("Left Speed: %d, Right Speed: %d\n", leftSpeed, rightSpeed);
+    printf("Left Speed: %.2f, Right Speed: %.2f\n", leftSpeed, rightSpeed);
     // Skriver ut states, beroende på x och y (for testing)
     switch (dir) {
         case FORWARD:
